+++ iblrig/base_tasks.py
@@ -39,6 +39,9 @@ import iblrig.graphic as graph
 import ibllib.io.session_params as ses_params
 from iblrig.transfer_experiments import BehaviorCopier
 
+# if HAS_PYSPIN:
+#     import PySpin
+
 OSC_CLIENT_IP = "127.0.0.1"
 
 
@@ -86,13 +89,10 @@ class BaseSession(ABC):
         self.iblrig_settings = iblrig.path_helper.load_settings_yaml(file_iblrig_settings or 'iblrig_settings.yaml')
         if iblrig_settings is not None:
             self.iblrig_settings.update(iblrig_settings)
-        if self.iblrig_settings['iblrig_local_data_path'] is None:
-            self.iblrig_settings['iblrig_local_data_path'] = Path.home().joinpath('iblrig_data')
-        else:
-            self.iblrig_settings['iblrig_local_data_path'] = Path(self.iblrig_settings['iblrig_local_data_path'])
         # Load the tasks settings, from the task folder or override with the input argument
         task_parameter_file = task_parameter_file or Path(inspect.getfile(self.__class__)).parent.joinpath('task_parameters.yaml')
         self.task_params = Bunch({})
+        self.wizard = wizard
 
         # first loads the base parameters for a given task
         if self.base_parameters_file is not None and self.base_parameters_file.exists():
@@ -150,17 +150,18 @@ class BaseSession(ABC):
         DATA_FILE_PATH: contains the bpod trials
             >>> C:\iblrigv8_data\mainenlab\Subjects\SWC_043\2019-01-01\001\raw_task_data_00\_iblrig_taskData.raw.jsonable  # noqa
         """
+        rig_computer_paths = iblrig.path_helper.get_local_and_remote_paths(
+            local_path=self.iblrig_settings['iblrig_local_data_path'],
+            remote_path=self.iblrig_settings['iblrig_remote_data_path'],
+            lab=self.iblrig_settings['ALYX_LAB']
+        )
         paths = Bunch({'IBLRIG_FOLDER': Path(iblrig.__file__).parents[1]})
         paths.BONSAI = paths.IBLRIG_FOLDER.joinpath('Bonsai', 'Bonsai.exe')
         paths.VISUAL_STIM_FOLDER = paths.IBLRIG_FOLDER.joinpath('visual_stim')
-        paths.LOCAL_SUBJECT_FOLDER = self.iblrig_settings['iblrig_local_data_path'].joinpath(
-            self.iblrig_settings['ALYX_LAB'] or '', 'Subjects')
-        paths.REMOTE_SUBJECT_FOLDER = (Path(self.iblrig_settings['iblrig_remote_data_path']).joinpath('Subjects')
-                                       if self.iblrig_settings['iblrig_remote_data_path'] else None)
+        paths.LOCAL_SUBJECT_FOLDER = rig_computer_paths['local_subjects_folder']
+        paths.REMOTE_SUBJECT_FOLDER = rig_computer_paths['remote_subjects_folder']
         # initialize the session path
-        date_folder = self.iblrig_settings['iblrig_local_data_path'].joinpath(
-            self.iblrig_settings['ALYX_LAB'] or '',
-            'Subjects',
+        date_folder = paths.LOCAL_SUBJECT_FOLDER.joinpath(
             self.session_info.SUBJECT_NAME,
             self.session_info.SESSION_START_TIME[:10],
         )
@@ -386,7 +387,7 @@ class BaseSession(ABC):
         self.logger.critical("Graceful exit")
         self.logger.info(f'Session {self.paths.SESSION_RAW_DATA_FOLDER}')
         self.session_info.SESSION_END_TIME = datetime.datetime.now().isoformat()
-        if self.interactive:
+        if self.interactive and not self.wizard:
             self.session_info.POOP_COUNT = graph.numinput(
                 "Poop count", f"{self.session_info.SUBJECT_NAME} droppings count:", nullable=True, askint=True)
         self.save_task_parameters_to_json_file()
@@ -447,6 +448,9 @@ class OSCClient(udp_client.SimpleUDPClient):
 
     def __init__(self, port, ip="127.0.0.1"):
         super(OSCClient, self).__init__(ip, port)
+
+    def __del__(self):
+        self._sock.close()
 
     def send2bonsai(self, **kwargs):
         """
@@ -520,10 +524,17 @@ class BonsaiRecordingMixin(object):
         desired borders of rig features, the actual triggering of the  cameras is done in the trigger_bonsai_cameras method.
         """
 
-        # TODO: spinnaker SDK
-
         if self._camera_mixin_bonsai_get_workflow_file(self.hardware_settings.get('device_cameras', None)) is None:
             return
+
+        # # TODO
+        # # enable trigger mode - if PySpin is available
+        # if HAS_PYSPIN:
+        #     pyspin_system = PySpin.System.GetInstance()
+        #     pyspin_cameras = pyspin_system.GetCameras()
+        #     for cam in pyspin_cameras:
+        #         cam.Init()
+        #         cam.TriggerMode.SetValue(True)
 
         bonsai_camera_file = self.paths.IBLRIG_FOLDER.joinpath('devices', 'camera_setup', 'setup_video.bonsai')
         # this locks until Bonsai closes
@@ -747,25 +758,6 @@ class RotaryEncoderMixin:
 
 
 class ValveMixin:
-    def get_session_reward_amount(self: object) -> float:
-        # simply returns the reward amount if no adaptive rewared is used
-        if not self.task_params.ADAPTIVE_REWARD:
-            return self.task_params.REWARD_AMOUNT
-        # simply returns the reward amount if no adaptive rewared is used
-        if not self.task_params.ADAPTIVE_REWARD:
-            return self.task_params.REWARD_AMOUNT
-        else:
-            raise NotImplementedError
-        # todo: training choice world reward from session to session
-        # first session : AR_INIT_VALUE, return
-        # if total_water_session < (subject_weight / 25):
-        #   minimum(last_reward + AR_STEP, AR_MAX_VALUE)  3 microliters AR_MAX_VALUE
-        # last ntrials strictly below 200:
-        #   keep the same reward
-        # trial between 200 and above:
-        #   maximum(last_reward - AR_STEP, AR_MIN_VALUE)  1.5 microliters AR_MIN_VALUE
-
-        # when implementing this make sure the test is solid
 
     def init_mixin_valve(self: object):
         self.valve = Bunch({})