+++ iblrig/test/test_transfers.py
@@ -3,6 +3,7 @@ from pathlib import Path
 import random
 import tempfile
 import unittest
+from unittest import mock
 
 from ibllib.io import session_params
 
@@ -57,11 +58,21 @@ class TestIntegrationTransferExperiments(unittest.TestCase):
             with tempfile.TemporaryDirectory(ignore_cleanup_errors=True) as td:
                 session = _create_behavior_session(td, ntrials=50, hard_crash=hard_crash)
                 session.paths.SESSION_FOLDER.joinpath('transfer_me.flag').touch()
-                iblrig.commands.transfer_data(local_subjects_path=session.paths.LOCAL_SUBJECT_FOLDER,
-                                              remote_subjects_path=session.paths.REMOTE_SUBJECT_FOLDER)
+                iblrig.commands.transfer_data(local_path=session.iblrig_settings['iblrig_local_data_path'],
+                                              remote_path=session.iblrig_settings['iblrig_remote_data_path'])
                 sc = BehaviorCopier(session_path=session.paths.SESSION_FOLDER,
                                     remote_subjects_folder=session.paths.REMOTE_SUBJECT_FOLDER)
                 self.assertEqual(sc.state, 3)
+
+        # Check that the settings file is used when no path passed
+        with tempfile.TemporaryDirectory(ignore_cleanup_errors=True) as td:
+            session = _create_behavior_session(td, ntrials=50, hard_crash=hard_crash)
+            session.paths.SESSION_FOLDER.joinpath('transfer_me.flag').touch()
+            with mock.patch('iblrig.path_helper.load_settings_yaml', return_value=session.iblrig_settings):
+                iblrig.commands.transfer_data()
+            sc = BehaviorCopier(session_path=session.paths.SESSION_FOLDER,
+                                remote_subjects_folder=session.paths.REMOTE_SUBJECT_FOLDER)
+            self.assertEqual(sc.state, 3)
 
     def test_behavior_do_not_copy_dummy_sessions(self):
         """
@@ -73,10 +84,8 @@ class TestIntegrationTransferExperiments(unittest.TestCase):
             with tempfile.TemporaryDirectory() as td:
                 session = _create_behavior_session(td, ntrials=ntrials)
                 session.paths.SESSION_FOLDER.joinpath('transfer_me.flag').touch()
-                iblrig.commands.transfer_data(
-                    local_subjects_path=session.paths.LOCAL_SUBJECT_FOLDER,
-                    remote_subjects_path=session.paths.REMOTE_SUBJECT_FOLDER
-                )
+                iblrig.commands.transfer_data(local_path=session.iblrig_settings['iblrig_local_data_path'],
+                                              remote_path=session.iblrig_settings['iblrig_remote_data_path'])
                 sc = BehaviorCopier(
                     session_path=session.paths.SESSION_FOLDER,
                     remote_subjects_folder=session.paths.REMOTE_SUBJECT_FOLDER)