+++ iblrig_tasks/_iblrig_tasks_trainingChoiceWorld/task.py
@@ -2,6 +2,7 @@ from iblrig.base_choice_world import TrainingChoiceWorldSession
 import iblrig.misc
 
 TRAINING_PHASE = -1
+ADAPTIVE_REWARD = -1.0
 
 
 class Session(TrainingChoiceWorldSession):
@@ -12,7 +13,11 @@ class Session(TrainingChoiceWorldSession):
         """ :return: argparse.parser() """
         parser = super(Session, Session).extra_parser()
         parser.add_argument('--training_phase', option_strings=['--training_phase'],
-                            dest='training_phase', default=TRAINING_PHASE, type=int)
+                            dest='training_phase', default=TRAINING_PHASE, type=int,
+                            help='defines the set of contrasts presented to the subject')
+        parser.add_argument('--adaptive_reward', option_strings=['--adaptive_reward'],
+                            dest='adaptive_reward', default=ADAPTIVE_REWARD, type=float,
+                            help='reward volume in microliters')
         return parser